use InteractiveConsole
